requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
Pillow==10.1.0
qrcode==7.4.2
python-dotenv==1.0.0